WIP controller refactoring for isolation
