+++ config/routes.rb
@@ -101,8 +101,8 @@ Rails.application.routes.draw do
     resources :event_notes, only: [:create, :update]
 
     member do
-      get '/v3' => 'students#show'
-      get :student_report
+      get '/v3' => 'profile#show_v3'
+      get '/student_report' => 'profile_pdf#student_report'
       get :restricted_notes
       get :photo
       post :service
